conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.2=py38h50d1736_1
argon2-cffi=20.1.0=py38h5406a74_2
async_generator=1.10=py_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.1=py_0
blas=1.0=mkl
bleach=3.3.0=pyh44b312d_0
blessings=1.7=py38_1002
bzip2=1.0.8=hc929b4f_4
c-ares=1.17.1=hc929b4f_0
ca-certificates=2021.5.30=h033912b_0
cairo=1.16.0=h4ec454a_1007
certifi=2021.5.30=py38h50d1736_0
cffi=1.14.5=py38ha97d567_0
cloudpickle=1.6.0=py_0
cycler=0.10.0=py38_0
cytoolz=0.11.0=py38haf1e3a3_0
dask-core=2.30.0=py_0
dataclasses=0.6=pypi_0
decorator=4.4.2=py_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
ffmpeg=4.3.1=he496704_1
fontconfig=2.13.1=h07ff6b0_1003
freetype=2.10.4=ha233b18_0
future=0.18.2=pypi_0
gettext=0.19.8.1=h7937167_1005
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.13=h2e338ed_1001
h5py=2.10.0=py38h0601b69_1
harfbuzz=2.7.4=hf28d55f_0
hdf5=1.10.6=nompi_hc5d9132_1113
icu=68.1=h74dc148_0
imageio=2.9.0=py_0
importlib-metadata=3.7.3=py38h50d1736_0
iniconfig=1.1.1=py_0
intel-openmp=2019.4=233
ipykernel=5.5.0=py38h9bb44b7_1
ipympl=0.7.0=pyhd8ed1ab_0
ipython=7.21.0=py38h9bb44b7_0
ipython_genutils=0.2.0=py_1
ipywidgets=7.6.3=pyhd3deb0d_0
jasper=1.900.1=h636a363_1006
jedi=0.18.0=py38h50d1736_2
jinja2=2.11.3=pyh44b312d_0
joblib=0.17.0=py_0
jpeg=9d=hbcb3906_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter_client=6.1.12=pyhd8ed1ab_0
jupyter_core=4.7.1=py38h50d1736_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
kiwisolver=1.3.0=py38h23ab428_0
kornia=0.6.1=pypi_0
kornia-moons=0.1.6=pypi_0
krb5=1.17.2=h60d9502_0
lame=3.100=h35c211d_1001
lcms2=2.11=h92f6f08_0
libblas=3.9.0=1_h508aa58_netlib
libcblas=3.9.0=3_h16b37ef_netlib
libcurl=7.71.1=h9bf37e3_8
libcxx=11.0.0=h4c3b8ed_1
libedit=3.1.20191231=h1de35cc_1
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=h6c81a4c_15
libgfortran5=9.3.0=h6c81a4c_15
libglib=2.66.4=h7424822_1
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=3_h16b37ef_netlib
liblapacke=3.9.0=3_h16b37ef_netlib
libnghttp2=1.41.0=h7580e61_2
libopencv=4.5.0=py38_4
libpng=1.6.37=ha441bb4_0
libprotobuf=3.13.0.1=hab81aa3_0
libsodium=1.0.18=hbcb3906_1
libssh2=1.9.0=h8a08a2b_5
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libwebp-base=1.1.0=hbcb3906_3
libxml2=2.9.10=h802ca4f_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.2=h79c402e_3
markupsafe=1.1.1=py38h5406a74_3
matplotlib-base=3.3.2=py38h181983e_0
mistune=0.8.4=py38h5406a74_1003
mit-semseg=1.0.0=pypi_0
mkl=2019.4=233
mkl-service=2.3.0=py38hfbe908c_0
mkl_fft=1.2.0=py38hc64f4ea_0
mkl_random=1.1.1=py38h959d312_0
more-itertools=8.5.0=py_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py38h50d1736_3
nbformat=5.1.2=pyhd8ed1ab_1
ncurses=6.2=h0a44026_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
nettle=3.6=hedd7734_0
networkx=2.5=py_0
ninja=1.10.1=py38h879752b_0
notebook=6.3.0=py38hecd8cb5_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
opencv=4.5.0=py38_4
opencv-python=4.5.4.60=pypi_0
opencv-python-headless=4.5.4.58=pypi_0
openh264=2.1.1=hfd3ada9_0
openssl=1.1.1k=h0d85af4_0
packaging=20.4=py_0
pandoc=2.13=h0d85af4_0
pandocfilters=1.4.2=py_1
parso=0.8.1=pyhd8ed1ab_0
path=15.0.0=py38_0
path.py=12.5.0=0
pcre=8.44=hb1e8313_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.0.1=py38h5270095_0
pip=21.0.1=py38hecd8cb5_0
pixman=0.40.0=hbcb3906_0
pluggy=0.13.1=py38_0
progressbar2=3.37.1=py38_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.18=pyha770c72_0
protobuf=3.13.0.1=py38hb1e8313_1
ptyprocess=0.7.0=pyhd3deb0d_0
py=1.9.0=py_0
py-opencv=4.5.0=py38h9bb44b7_4
pycparser=2.20=pyh9f0ad1d_2
pydegensac=0.1.2=pypi_0
pygments=2.8.1=pyhd8ed1ab_0
pyparsing=2.4.7=py_0
pyrsistent=0.17.3=py38h5406a74_2
pytest=6.1.1=py38_0
pytest-runner=5.2=py_0
python=3.8.5=h26836e1_1
python-dateutil=2.8.1=py_0
python-utils=2.4.0=py38_0
python_abi=3.8=1_cp38
pywavelets=1.1.1=py38haf1e3a3_2
pyyaml=5.3.1=py38haf1e3a3_1
pyzmq=19.0.2=py38h2c785a9_2
readline=8.0=h1de35cc_0
scikit-image=0.17.2=py38h81aa140_0
scikit-learn=0.24.1=py38hb2f4e1b_0
scipy=1.6.0=py38h95ea19e_0
send2trash=1.5.0=py_0
setuptools=50.3.1=py38hecd8cb5_1
six=1.15.0=py_0
sqlite=3.33.0=hffcf06c_0
tensorboardx=2.1=py_0
terminado=0.9.3=py38h50d1736_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=py38h0cf3a3e_2
tk=8.6.10=hb0a8c7a_0
toml=0.10.1=py_0
toolz=0.11.1=py_0
torch=1.10.0=pypi_0
torchvision=0.8.1=py38_cpu
tornado=6.0.4=py38h1de35cc_1
tqdm=4.50.2=py_0
traitlets=5.0.5=py_0
typing_extensions=3.7.4.3=py_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=py_1
wheel=0.35.1=py_0
widgetsnbextension=3.5.1=py38h50d1736_4
x264=1!152.20180806=h1de35cc_0
xz=5.2.5=h1de35cc_0
yacs=0.1.8=pypi_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.5=h41d2c2f_0
